// src/app/core/enums/ui/modals/modalNamesEnum.ts
/**
 * @description Modal 列表 (enum)
 */
export enum ModalNamesEnum {
    DataStorageModal = 'dataStorageModal',
    SubmissionModal = 'submissionModal',
    SubmissionHintModal = 'submissionHintModal',
    SuccessSubmissionModal = 'successSubmissionModal',
    FailSubmissionModal = 'failSubmissionModal',
    PreviewModal = 'previewModal',
    PreviewMsgModal = 'previewMsgModal',
    OnlineAuthApplyModal = 'onlineAuthApplyModal',
    HealthDeclarationForm = 'HealthDeclarationForm', // 健康告知書（被保人）
    InsuranceRightsConfirmation = 'InsuranceRightsConfirmation', // 投保權益確認書
    ReviewPeriodConfirmation = 'ReviewPeriodConfirmation', // 審閱期間確認聲明書
    FinancialDisclosure = 'FinancialDisclosure', // 財務狀況告知書
    DataUsageNotice = 'DataUsageNotice', // 個人資料蒐集、處理、利用告知事項
    ApplicationInstructions = 'ApplicationInstructions', // 要保書填寫說明
    PolicyholderInfo = 'PolicyholderInfo', // 人壽保險投保人須知
    EPolicyApplication = 'EPolicyApplication', // 電子保單申請書
    HealthQuestionnaire = 'HealthQuestionnaire', // 疾病問券
    CRSForm = 'CRSForm', // CRS自我證明表
    AgentDisclosure = 'AgentDisclosure', // 業告書
    GeneralDiseaseQuestionnaire = 'GeneralDiseaseQuestionnaire', // 一般疾病問卷
    ThyroidDiseaseQuestionnaire = 'ThyroidDiseaseQuestionnaire', // 甲狀腺疾病問卷
    HepatitisOrHepatitisCarrierQuestionnaire = 'HepatitisOrHepatitisCarrierQuestionnaire', // 肝炎或肝炎病毒帶原問卷
    UrinarySystemDiseaseQuestionnaire = 'UrinarySystemDiseaseQuestionnaire', // 泌尿系統疾病問卷
    EsophagusStomachQuestionnaire = 'EsophagusStomachQuestionnaire', // 食道、胃或十二指腸潰瘍問卷
    BronchialDiseaseQuestionnaire = 'BronchialDiseaseQuestionnaire', // 氣喘/慢性支氣管炎/慢性阻塞性肺疾病問卷
    HypertensionQuestionnaire = 'HypertensionQuestionnaire', // 高血壓問卷
    HyperuricemiaQuestionnaire = 'HyperuricemiaQuestionnaire', // 高尿酸血症/痛風/關節炎問卷
    DiabetesQuestionnaire = 'DiabetesQuestionnaire', // 糖尿病問卷
    EpilepsyQuestionnaire = 'EpilepsyQuestionnaire', // 癲癇問卷
    HeartDiseaseQuestionnaire = 'HeartDiseaseQuestionnaire', // 心臟疾病問卷
    TumorAndPolypQuestionnaire = 'TumorAndPolypQuestionnaire', // 腫瘤及息肉問卷
    PoliomyelitisQuestionnaire = 'PoliomyelitisQuestionnaire', // 小兒麻痹問卷
    TraumaAndSkeletalInjuryQuestionnaire = 'TraumaAndSkeletalInjuryQuestionnaire', // 外傷骨骼問卷
    EyeDiseaseQuestionnaire = 'EyeDiseaseQuestionnaire' // 眼部疾病問卷
};